SnsPostPostingPage 컴포넌트의 state, method들을 CreatePost 컴포넌트로 이동
